chore: Remove unused files and refactor file structure
