--- client/src/components/chat/chat.css
@@ -1,168 +1,0 @@
-.chat{
-    flex: 2;
-    border-right: 1px solid #e8e2e2;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-
-    .top{
-        padding: 10px;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        border-bottom: 1px solid #e8e2e2;
-
-        .user{
-            display: flex;
-            align-items: center;
-            gap: 20px;
-
-            img{
-                width: 50px;
-                height: 50px;
-                border-radius: 50%;
-                object-fit: cover;
-            }
-
-            .texts{
-                display: flex;
-                flex-direction: column;
-                gap: 2px;
-
-                span{
-                    font-size:16px;
-                    font-weight: bold;
-                }
-                p{
-                    font-size: 14px;
-                    font-weight: 300;
-                    color: #6a7087;
-                }
-            }
-        }
-        .icons{
-            padding-right: 5px;
-            display: flex;
-            gap: 16px;
-        
-            img{
-                width: 20px;
-                height: 20px;
-                
-            }
-        }
-
-    }
-
-
-    .center{
-        padding: 20px; 
-        flex: 1;
-        overflow: scroll;
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-        padding-bottom: 60px; 
-    
-        .message{
-            max-width: 70%;
-            display: flex;
-            gap: 15px;
-        
-        &.own{
-            align-self: flex-end;
-    
-            .texts{
-                p{
-                    background-color: #882A85;
-    
-                }
-            }
-    
-        }
-        img{
-            width: 30px;
-            height: 30px;
-            border-radius: 50%;
-            object-fit: cover;
-        }
-        .texts{
-            flex: 1;
-            display: flex;
-            flex-direction: column;
-            gap: 5px;
-    
-            p{
-                padding: 15px;
-                background-color:#1f163d08;
-                border-radius: 10px;
-    
-            }
-            span{
-                font-size: 12px;
-            }
-    
-            }
-        }
-    }
-    
-    
-    .bottom{
-        position: absolute;
-        bottom: 0;
-        left: 0;
-        width: 100%;
-        padding: 10px;
-        display: flex;
-        background-color: white;
-        align-items: center;
-        justify-content: space-between;
-        border-top: 1px solid #e8e2e2;
-        gap: 20px;
-        margin-top: auto;
-        
-        
-        .icons{
-            display: flex;
-            gap: 20px;
-        }
-
-        img{
-            width: 20px;
-            height: 20px;
-            cursor: pointer;
-        }
-        .input{
-            flex: 1;
-            background-color: transparent;
-            border: none;
-            outline: none;
-            color: white;
-            padding: 20px;
-            border-radius: 10px;
-            font-size: 16px;
-
-
-        }
-
-        .emoji{
-            position: relative;
-
-            .picker{
-                position: absolute;
-                bottom: 50px;
-                left: 0;
-            }
-        }
-        .sendBotton{
-            background-color: #882A85;
-            color: white;
-            padding: 10px 20px;
-            border: none;
-            border-radius: 5px;
-            font-size: 15px;
-        }
-    }
-
-}
