--- client/src/components/list/chatList/chatlist.css
@@ -1,144 +1,0 @@
-.chatList {
-    flex: 1;
-    height: 100vh;
-    overflow-y: scroll;
-    background-color: #F1F2F6;
-    
-    
-    .header {
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        margin-bottom: 6px;
-    }
-    
-    .logoSection {
-        display: flex;
-        align-items: center;
-        width: 100%;
-
-
-    }
-    
-    .logoSection img {
-        width: 120px;
-        height: 120px;
-        object-fit: contain;
-    }
-
-    .settingsIcon{
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: 25px;           
-        height: 25px;
-        margin-left: 120px;
-        background-color: #f6cff5; 
-        border-radius: 10%;
-        cursor: pointer;
-    }
-    .settingsIcon img {
-        width: 15px;
-        height: 15px;
-        display: block;        
-    }
-    
-
-
-    .chatLine {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        margin-top: 10px;
-        gap: 1px;
-      }
-      
-      .addSymbol {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: 20px;
-        height: 20px;
-        background-color: #882A85;
-        border-radius: 50%;
-        margin-left: 12px;
-        margin-top: 6px;
-        cursor: pointer;
-
-      }
-      
-      .addSymbol img {
-        width: 10px;
-        height: 10px;
-        display: block;
-      }
-      
-      .chatTitle {
-        font-size: 30px;
-        font-weight: 700;
-        color: #333333;
-        margin-bottom: 5;
-        padding-left:20px ;
-      }
-
-    .search{
-        display: flex;
-        align-items: center;
-        gap: 20px;
-        padding: 20px;
-
-        .searchBar{
-            flex:1;
-            background-color: #a0a2aa44;
-            display: flex;
-            align-items: center;
-            gap: 20px;
-            border-radius: 20 px;
-            padding: 1px;
-
-            input{
-                background-color: transparent;
-                border: none;
-                outline: none;
-                color: black;
-                flex: 1;
-            
-            
-            }
-
-            img{
-                width: 12px;
-                height: 12px;
-            }
-        }
- 
-    }
-    .item{
-        display: flex;
-        align-items: center;
-        gap: 20px;
-        padding: 12px;
-        cursor: pointer;
-        border-bottom: 1px solid #e8e2e2;
-
-        img{
-            width: 45px;
-            height: 45px;
-            border-radius: 50%;
-            object-fit: cover;
-        }
-        .texts{
-            display: flex;
-            flex-direction: column;
-            gap: 5px;
-
-            span{
-                font-weight: 200;
-            }
-            p{
-                font-size: 12px;
-            }
-        }
-    }
-
-}
