--- client/src/components/list/list.css
@@ -1,5 +1,0 @@
-.list {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-}
