+++ client/src/components/sidePanel/settings/Settings.tsx
@@ -1,4 +1,4 @@
-import useAuth from "../../hooks/useAuth";
+import useAuth from "../../../hooks/useAuth";
 
 const Settings = () => {
   const { logout } = useAuth();
