+++ client/src/routes/(auth)/Register.tsx
