+++ client/src/routes/(auth)/components/Logo.tsx
