+++ client/src/routes/chatBoard/components/ChatFooter.tsx
