+++ client/src/routes/chatBoard/components/TypingIndicator.tsx
