--- client/src/routes/dashboard/page.tsx
@@ -1,5 +1,0 @@
-const Dashboard = () => {
-  return <div>Dashboard</div>;
-};
-
-export default Dashboard;
